♻️ FloatingButton 완성
FloatingButton 코드 간소화 및 애니메이션으로 UX 향상
